refactor: remove unused SCSS files and reorganize styles directory
