--- src/styles/Review/ReadPosts.scss
@@ -1,19 +1,0 @@
-// 리뷰 상세보기 박스 컴포넌트 디자인
-@use '../root.scss';
-
-.signal {
-  background: linear-gradient(145deg, #f0f0f0, #d9d9d9);
-  box-shadow: 0 1px 1px rgba(0, 0, 0, 0.2);
-
-  &.good {
-    background: linear-gradient(145deg, #4caf50, #388e3c);
-  }
-
-  &.neutral {
-    background: linear-gradient(145deg, #ffeb3b, #fbc02d);
-  }
-
-  &.bad {
-    background: linear-gradient(145deg, #f44336, #c62828);
-  }
-}
