+++ src/styles/components/EventBox.scss
